refactor:ipsec: move IPSecAgent definition to specific file
This commit moves the IPSecAgent definition to the ipsec_linux.go file.
This is to have everything related to the agent in one place, while leaving
the cell definition and the constructor returning the interface in the cell.go file.
This aligns to the WireGuard agent files structure. However, we do not
rename the file to preserve history.
